fix rtt type error

+++ kcp.go
@@ -228,6 +228,7 @@ func (kcp *KCP) Recv(buffer []byte) (n int) {
 			break
 		}
 	}
+
 	if count > 0 {
 		kcp.rcv_queue = kcp.rcv_queue[count:]
 	}
@@ -300,11 +301,11 @@ func (kcp *KCP) update_ack(rtt int32) {
 		kcp.rx_srtt = uint32(rtt)
 		kcp.rx_rttval = uint32(rtt) / 2
 	} else {
-		delta := uint32(rtt) - kcp.rx_srtt
+		delta := int(rtt) - int(kcp.rx_srtt)
 		if delta < 0 {
 			delta = -delta
 		}
-		kcp.rx_rttval = (3*kcp.rx_rttval + delta) / 4
+		kcp.rx_rttval = (3*kcp.rx_rttval + uint32(delta)) / 4
 		kcp.rx_srtt = (7*kcp.rx_srtt + uint32(rtt)) / 8
 		if kcp.rx_srtt < 1 {
 			kcp.rx_srtt = 1
